refactor(schema,compare): paga os 3 warnings de decomposicao do react-doctor
Fecha os ultimos warnings do bump 0.7.8; react-doctor vai a score 100.

no-many-boolean-props (ComparePage): as tres flags de estado de fila
resolvidas no servidor (showingAll, hasAssignedDocs, isImpersonating)
viraram um objeto queueContext. Descrevem a mesma coisa -- o contexto
da fila exibida -- e sao consumidas juntas na copy do estado vazio;
soltas, contavam como quatro props booleanas.

no-giant-component (SchemaEditor): decomposto de fato. Os quatro
handlers de acao de servidor (salvar, publicar MAJOR, reconstruir
historico, recuperar campos legados) sairam para um hook co-localizado
useSchemaEditorActions, no mesmo espirito de useSchemaDraft/
useSchemaEditorDialogs. Costura limpa, sem achado novo no fallow.

no-giant-component (ComparePage): SUPRIMIDO por arquivo, com ressalva
e rastreado em #564. E um componente reusavel, nao um route component,
entao a politica manda refatorar -- mas a decomposicao real (extrair a
orquestracao num controller) reduzia o cognitive de 89 para 42+29 e
passava em react-doctor/tsc/127 testes, porem esbarrava no gate
new-only do fallow: complexidade transplantada conta como introduzida
e um controller unico a cognitive 42 segue acima do limiar. Passar
exigiria fatiar em 3-4 sub-hooks focados, refactor de risco alto no
hot-path da Comparacao que nao cabe num PR de limpeza de lint. A
justificativa e a condicao de remocao do override estao em
docs/LINT_CONFIG.md.
