+++ docs/LINT_CONFIG.md
@@ -134,7 +134,8 @@ O re-scan da 0.5.8 (2026-06-23) deixou 36 diagnósticos da categoria **Bugs** se
 Fase 0 do épico #239 (rastreada por #339): PR de baixo risco que remove 5 warnings baratos da 0.5.8 **sem decompor componentes**. Um foi resolvido por deleção de causa raiz, os outros 4 por supressão auditada.
 
 - **`deslop/unused-file`** em `src/lib/supabase/client.ts`: **deletado, não suprimido.** Órfão (zero imports em `src`); a única export `createBrowserClient` usava o template JWT antigo `getToken({ template: "supabase" })`, morto após a migração de auth RS256 (#299/#309). Fix de causa raiz.
-- **`no-giant-component`** em `src/app/(app)/projects/[id]/analyze/compare/page.tsx:59` (`ComparePageRoute`, 491 linhas) e `src/app/(app)/projects/[id]/reviews/comments/page.tsx:8` (`CommentsPage`, 433 linhas): override **por arquivo**, escopado só a estes 2 **page-level route components** (a #339 endossa suprimir esses — são fronteiras de rota, não unidades reusáveis a decompor). A regra segue ativa nos outros 2 disparos (`LotteryDialog.tsx:201`, `DocumentsPageClient.tsx:70`), que **não** são route components e ficam para refactor real numa fase posterior — desligar a regra global os mascararia.
+- **`no-giant-component`** em `src/app/(app)/projects/[id]/analyze/compare/page.tsx:59` (`ComparePageRoute`, 491 linhas) e `src/app/(app)/projects/[id]/reviews/comments/page.tsx:8` (`CommentsPage`, 433 linhas): override **por arquivo**, escopado só a estes 2 **page-level route components** (a #339 endossa suprimir esses — são fronteiras de rota, não unidades reusáveis a decompor). A regra segue ativa nos disparos que **não** são route components (`LotteryDialog.tsx`, `DocumentsPageClient.tsx`), que ficam para refactor real numa fase posterior — desligar a regra global os mascararia.
+- **`no-giant-component`** em `src/components/compare/ComparePage.tsx` (497 linhas, cognitive 89): override **por arquivo**, com a ressalva de que **este é um componente reusável, não um route component** — a política acima manda refatorar, não suprimir. A supressão é **temporária e rastreada em #564**. No bump 0.7.8 tentei a decomposição de fato (extrair a orquestração num hook `useComparePageController`): reduzia o máximo de cognitive de 89 para 42+29, passava em react-doctor, typecheck e nos 127 testes de `compare`, mas **esbarrava no gate `new-only` do fallow** — a complexidade transplantada conta como introduzida e um único controller a cognitive 42 segue acima do limiar. Passar exigiria fatiar o controller em 3–4 sub-hooks focados, refactor de risco alto no hot-path da Comparação que não cabia num PR de limpeza de lint feito sem supervisão. O `SchemaEditor`, o outro `no-giant-component` do mesmo bump, **foi decomposto** (costura limpa: 4 handlers de ação → `useSchemaEditorActions`, sem achado novo no fallow) — por isso não aparece aqui. Ao decompor o `ComparePage` (via #564), remover este override.
 - **`no-json-parse-stringify-clone`** em `src/lib/__tests__/lottery-utils.test.ts:518`: supressão **inline**. Deep clone (`JSON.parse(JSON.stringify(...))`) de uma fixture plana de `coOccurrence` para snapshot do teste de imutabilidade — FP legítimo em arquivo de teste; o objeto não tem `Date`/`Map`/`undefined` que o round-trip perderia.
 - **`async-defer-await`** em `src/components/shared/RunLlmButton.tsx:62`: supressão **inline**. O `await requireSupabaseToken(getToken)` roda antes do guard de cancelamento **de propósito** — o token do template expira em ~60s e precisa estar fresco quando a request de status parte a cada poll; mover o await para baixo do guard atrasaria a renovação. Mesma razão da supressão inline que já existia logo abaixo, no `await fetchFastAPI` do mesmo `poll` (esta é a 2ª ocorrência da regra no mesmo bloco; diagnóstico novo, não previsto na #339).
