+++ frontend/src/components/compare/ComparePage.tsx
@@ -57,10 +57,17 @@ interface ComparePageProps {
   // "Meus atribuídos" / "Todos" — coordenador também compara documentos, por
   // isso o padrão é a fila pessoal dele, igual pesquisador.
   isCoordinator: boolean;
-  // Valor efetivo (resolvido no servidor) da aba de fila atual — fonte única
-  // pro valor exibido em CompareQueueTabs (evita reler a URL no cliente) e
-  // pra mensagem do estado vazio.
-  showingAllQueue: boolean;
+  // Estado da fila e da sessão, todo resolvido no servidor. Agrupado num objeto
+  // porque as três flags descrevem a mesma coisa — o contexto da fila exibida —
+  // e são consumidas juntas na copy do estado vazio; soltas, contavam como
+  // quatro props booleanas e disparavam `no-many-boolean-props`.
+  queueContext: QueueContext;
+}
+
+interface QueueContext {
+  // Valor efetivo da aba de fila atual — fonte única pro valor exibido em
+  // CompareQueueTabs (evita reler a URL no cliente) e pra mensagem do vazio.
+  showingAll: boolean;
   // Se o coordenador TEM documentos atribuídos a ele para comparação (mesmo
   // que nenhum tenha passado nos filtros de cobertura/divergência) — usado só
   // pra diferenciar a mensagem de estado vazio: "sem nada atribuído" (trocar
@@ -94,10 +101,10 @@ export function ComparePage({
   currentUserId,
   canManageAnyPair,
   isCoordinator,
-  showingAllQueue,
-  hasAssignedDocs,
-  isImpersonating,
+  queueContext,
 }: ComparePageProps) {
+  const { showingAll: showingAllQueue, hasAssignedDocs, isImpersonating } =
+    queueContext;
   // Impersonação master torna a Comparação somente-leitura (issue #428). Mesma
   // convenção `readOnly: boolean` do Codificar (`code/page.tsx`).
   const readOnly = isImpersonating;
